Added joystick integration in the Selection menu. FixedThrall Red skin.

+++ source/Terrrain.cpp
@@ -17,7 +17,7 @@ void Terrain::generateLevel(unsigned int seed)
   std::vector<Vect<2u, unsigned int>> rooms;
 
   for (unsigned int i(0); i < 30; ++i)
-    rooms.emplace_back(rangeX(engine), rangeY(engine));
+    rooms.emplace_back(static_cast<unsigned int>(rangeX(engine)), static_cast<unsigned int>(rangeY(engine)));
   std::vector<Vect<2u, unsigned int>> startNodes{};
 
   rooms.emplace_back(10u, 10u);
@@ -27,8 +27,8 @@ void Terrain::generateLevel(unsigned int seed)
     {
       bool genConnection(true);
 
-      Vect<2u, unsigned int> size((range10(engine) + range10(engine)) / 2, (range10(engine) + range10(engine)) / 2);
-      Vect<2u, unsigned int> relative(range5(engine), range5(engine));
+      Vect<2u, unsigned int> size(static_cast<unsigned int>((range10(engine) + range10(engine)) / 2), static_cast<unsigned int>((range10(engine) + range10(engine)) / 2));
+      Vect<2u, unsigned int> relative(static_cast<unsigned int>(range5(engine)), static_cast<unsigned int>(range5(engine)));
 
       for (Vect<2u, unsigned int> i(0u, 0u); i[1] != size[1]; ++i[1])
 	for (i[0] = 0u; i[0] != size[0]; ++i[0])
